Full page thesis captions now appear on own page
Thanks to eddie schlafly for help on this one. I ended up not using his patch exactly, because that seemed to cause problems. Instead I merely added a \clearpage command at hte begining of the create FP figure and I Changed one p! to a b!.

Next step is to add a caption header to the figure.