It works. Basically
This does everything right. It makes the PDF hyperlink correct. It makes a second that is numbered correctly. The only annoying thing is if there is a second small float, liek as mall table, that can end up on the caption page. So I need to set a parameter taht says there shoudl only be one float per page. I can do that and that is what I will work on now.